+++ The game/Assets/CharacterControl.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f7a7f6063cd020843aef3135ce359ac5
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
