+++ The game/Assets/Scripts/CharactreController2D.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f570c5bfdf7810248a4e0ff46f57da02
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
