+++ The game/Assets/Sprites/Knight/The knight.controller
@@ -146,7 +146,7 @@ AnimatorStateMachine:
     m_Position: {x: 120, y: 200, z: 0}
   - serializedVersion: 1
     m_State: {fileID: 6603835562820243519}
-    m_Position: {x: 235, y: 65, z: 0}
+    m_Position: {x: 280, y: 120, z: 0}
   m_ChildStateMachines: []
   m_AnyStateTransitions: []
   m_EntryTransitions: []
